services/auth-service/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
grpcio>=1.60.0
grpcio-health-checking>=1.60.0
psycopg2-binary>=2.9
python-decouple>=3.8
